Healthbar visual bug fixed
current health was being calculated by subtracting 20 * time.deltatime from the current health, but only if currentheatlh was more than 0.1 away from targethealth.
That was the problem, the check isn't if currenthealth is 0.1 BIGGER, but if currentHealth is 0.1 AWAY from targetHealth.

This could cause a problem when The change in current health (currennthealth * time.deltatime) was larger than "currenthealth -targethealth". normally Time.Delta time would be very small, and this wouldn't happend, but if the game lagged or paused, this would happen.

Solution:
Clamp the amount that is subrtracted from currentHealth to be at max the difference between targetHealth and CurrentHealth.